multiplex

+++ MultiplexBookingSystem/src/main/java/com/capg/service/BookingServiceImpl.java
@@ -26,7 +26,6 @@ public class BookingServiceImpl implements BookingService {
     public List<Booking> findAll() {
         return bookingRepository.findAll();
     }
-
    
     @Override
     public Booking findById(long id) throws BookingNotFoundException{
@@ -39,9 +38,6 @@ public class BookingServiceImpl implements BookingService {
     }
     return book;
     }
-//    public Booking findById(long id) {
-//        return bookingRepository.findById(id).orElse(null);
-//    }
 
     @Override
     public List<Booking> findByDate(LocalDate date) {
@@ -49,74 +45,15 @@ public class BookingServiceImpl implements BookingService {
     }
 
     @Override
-    public void deleteBooking(Booking booking) {
-        bookingRepository.delete(booking);
+    public Booking deleteBooking(long id) {
+       Booking booking =bookingRepository.findById(id).orElse(null);
+       if(booking !=null) {
+    	   bookingRepository.delete(booking);
+       }
+       return booking;
     }
 }
 
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-//package com.capg.service;
-//
-//import java.time.LocalDate;
-//
-//import java.util.List;
-//
-//import javax.persistence.EntityManager;
-//import javax.persistence.PersistenceContext;
-//import javax.persistence.Query;
-//import javax.transaction.Transactional;
-//
-//import org.springframework.stereotype.Repository;
-//
-//import com.capg.entity.Booking;
-//
-//@Repository
-//@Transactional
-//public class BookingServiceImpl implements BookingService {
-//
-//	@PersistenceContext
-//	EntityManager entityManager;
-//	
-//	@Override
-//	public Booking addBooking(Booking booking) {
-//		entityManager.persist(booking);
-//		return booking;
-//	}
-//
-//	@Override
-//	public List<Booking> findAll() {
-//		Query query = entityManager.createQuery("SELECT b FROM Booking b");
-//		return query.getResultList();
-//	}
-//
-//	@Override
-//	public Booking findById(long id) {
-//		return entityManager.find(Booking.class, id);
-//	}
-//
-//	@Override
-//	public List<Booking> findByDate(LocalDate date) {
-//		Query query = entityManager.createQuery("SELECT b FROM Booking b WHERE b.bookedDate = :date");
-//		query.setParameter("date", date);
-//		return query.getResultList();
-//	}
-//
-//	@Override
-//	public void deleteBooking(Booking booking) {
-//		entityManager.remove(booking);
-//	}
-//
-//}
